got rid of some bugs

+++ python/tvm/relay/op/nn/nn.py
@@ -1176,11 +1176,11 @@ def upsampling(data,
 
     if isinstance(data, Expr) or isinstance(scale_h, Expr) or isinstance(scale_w, Expr):
         if not isinstance(data, Expr):
-            data = const(list(data), "int64") # is this right
+            data = const(list(data), "int32")
         if not isinstance(scale_h, Expr):
-            scale_h = const([scale_h], "int64")
+            scale_h = const([scale_h], "int32")
         if not isinstance(scale_w, Expr):
-            scale_w = const([scale_w], "int64")
+            scale_w = const([scale_w], "int32")
         return _dyn_make.upsampling(data, scale_h, scale_w, layout, method, align_corners)
     else: 
         return _make.upsampling(data, scale_h, scale_w, layout, method, align_corners)
